grouping_service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.0
torch==2.3.0
torchvision==0.18.0
